refactor: divider 분리
